[FIX] PhraseFinder architecture and Data structures using dataclass
